Move validation of `BloodSugarReading` into the model
